% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparingImpMethods.R
\name{CalcCorrAvg}
\alias{CalcCorrAvg}
\title{Calculates Correlation Ignoring Heterogeneity}
\usage{
CalcCorrAvg(corr_array_obs, num_of_metabos)
}
\arguments{
\item{corr_array_obs}{Output from CreateCorrelation()}

\item{num_of_metabos}{Number of metabolites}
}
\value{
Correlation matrix with dimension num_of_metabo by num_of_metabo, ignores heterogeneity
}
\description{
Calculates Correlation Ignoring Heterogeneity
}
