% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparingImpMethods.R
\name{CalculateR2}
\alias{CalculateR2}
\title{Calculates R2 across factors for different imputation types}
\usage{
CalculateR2(imputed.df, factor_mat)
}
\arguments{
\item{imputed.df}{dataframe of imputed values. Output from ImputeMissingMetabolites()}

\item{factor_mat}{Matrix where 3 columns are three factors and rows are values of each factor}
}
\value{
data frame of R2 values by imputation type
}
\description{
Creates data frame of R2 values across factors and imputation type. May take time with large sample size.
}
