% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparingImpMethods.R
\name{CreateAutoRegressiveMat}
\alias{CreateAutoRegressiveMat}
\title{Creates Autoregressive Correlation Matrix}
\usage{
CreateAutoRegressiveMat(num_of_metabos, rho, diag_val)
}
\arguments{
\item{num_of_metabos}{Number of metabolites (dimension of matrix)}

\item{rho}{Degree of autoregressiveness}

\item{diag_val}{Diagonal value. Set to 1 if m_0, 0 else}
}
\value{
Autoregressive square matrix of dimension num_of_metabo
}
\description{
Creates autoregressive correlation matrix of variable size with variable rho and diagonal values
}
