% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparingImpMethods.R
\name{CreateBlockMatrix}
\alias{CreateBlockMatrix}
\title{Creates Block Diagonal Correlation Matrix}
\usage{
CreateBlockMatrix(
  num_of_metabos,
  low_val,
  high_val,
  diag_val,
  size_of_block = 10
)
}
\arguments{
\item{num_of_metabos}{Number of metabolites (dimension of matrix)}

\item{low_val}{lower bound for block correlation. Correlation is uniform between low_val and high_val}

\item{high_val}{upper bound for block correlation. Correlation is uniform between low_val and high_val}

\item{diag_val}{Diagonal value. Set to 1 if m_0, 0 else}

\item{size_of_block}{Block size. Default is 10. num_of_metabos must be divisible by size_of_block}
}
\value{
Block diagonal matrix of dimension num_of_metabo
}
\description{
Creates block diagonal correlation matrix of variable size with blocks and diagonal values.
Correlation values are drawn from a random uniform distribution between low_val and high_val
}
