% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparingImpMethods.R
\name{CreateCorrelation}
\alias{CreateCorrelation}
\title{Creates true data, obs data, true correlation, and obs correlation}
\usage{
CreateCorrelation(
  m_0,
  m_1,
  m_2,
  m_3,
  num_of_metabos = 20,
  sample_size = 1000,
  num_to_remove = 5,
  factor_mat
)
}
\arguments{
\item{m_0}{Residual correlation matrix}

\item{m_1}{Correlation matrix for factor 1}

\item{m_2}{Correlation matrix for factor 2}

\item{m_3}{Correlation matrix for factor 3}

\item{num_of_metabos}{Number of metabolites}

\item{sample_size}{Number of individuals generated from correlation matrices}

\item{num_to_remove}{Number of metabolites to remove. Must be smaller than 1/4*num_of_metabolites}

\item{factor_mat}{Matrix where 3 columns are three factors and rows are values of each factor}
}
\value{
true data, obs data, true correlation, and obs correlation
}
\description{
First step of imputation pipeline. Correlation structure is m_0 + (m_1xF_1) + (m_2xF_2) + (m_3xF_3)
factor_mat is a matrix which contains the values for the 3 different factors. Each column corresponds to a factor and
each row is a value. By setting the third column (factor) to zero, we can generate data from a 2 factor model instead of 3.
Suppose that factor_mat has 30 rows (30 different combinations of factors), then 120 (30 x 4) data sets are generated in total.
num_to_remove metabolites are removed in four increments. Set num_to_remove=5 and that the first row of factor mat is a, b, c.
Under the same factors, in the first data set metabolites 1:5 are removed, in the second data set 6:10 are removed, in the third
11:15, and in the fourth 16:20. This then repeats with the second row of factor mat.
}
