% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparingImpMethods.R
\name{FactorAnalysis}
\alias{FactorAnalysis}
\title{Runs factor analysis}
\usage{
FactorAnalysis(reorganized_corr_mice_med, num_of_metabos = 20)
}
\arguments{
\item{reorganized_corr_mice_med}{Second output from ReorganizeAndMICE(). Fischer transformed 2d matrix}

\item{num_of_metabos}{Number of metabolites}
}
\value{
list of lists. First list is factor analysis intermediates for 1-4 factors, second is 3d corr matrices for 1-4 factors
}
\description{
Runs factor analysis with 1-4 factors. Converts input from 2d to 3d form (such as in the description of ReogranizeAndMICE()).
Outputs both factor analysis intermediates (loadings, scores, ...) and estimated correlation for each study.
Input is second output from ReorganizeAndMICE(), where the columns are studies and the rows are pairwise correlation.
Runs factor analysis on transpose of input.
}
