% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparingImpMethods.R
\name{PlotR2OneFact}
\alias{PlotR2OneFact}
\title{Plots R2 of one factor imputation}
\usage{
PlotR2OneFact(r2.df)
}
\arguments{
\item{r2.df}{R2 dataframe, output from CalculateR2()}
}
\value{
Line chart of R2 values by imputation type
}
\description{
Plots R2 in line chart of different imputation methods.
Useful when underlying data is generated from a single factor.
}
