% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparingImpMethods.R
\name{SwarmGenerator}
\alias{SwarmGenerator}
\title{Generates Swarm File}
\usage{
SwarmGenerator(
  seed_max,
  swarm_name,
  r_name,
  low_val_1,
  high_val_1,
  low_val_2,
  high_val_2
)
}
\arguments{
\item{seed_max}{Max seed number. Runs this many simulations}

\item{swarm_name}{Name of swarm file output}

\item{r_name}{Name of r file to run}

\item{low_val_1}{low value for m_1}

\item{high_val_1}{high value for m_1}

\item{low_val_2}{low value for m_2}

\item{high_val_2}{high value for m_2}
}
\value{
writes out swarm file
}
\description{
Generates Swarm File
}
